+++ src/scripts/get-roots.js
@@ -21,8 +21,6 @@ const configRoot = process.platform === `win32`
 
 // base paths
 const roots = {
-	preBuild: path.join(moduleRoot, `.pre-build`),
-	moduleBuild: path.join(moduleRoot, `.build`),
 	dist: path.join(moduleRoot, `dist`),
 	src: path.join(moduleRoot, `src`),
 	eyasBuild: path.join(consumerRoot, `.eyas-preview`),